requirements.txt
aubio==0.4.9
classy_vision==0.6.0
config==0.5.1
librosa==0.9.2
matplotlib==3.3.3
numba==0.48.0
numpy==1.23.0
psutil==5.7.2
pyloudnorm==0.1.0
git+https://github.com/csteinmetz1/pymixconsole
pypesq==1.2.4
pytorch_lightning==1.3.2
PyYAML==5.4
scikit_learn==1.1.3
scipy==1.6
SoundFile==0.10.3.post1
soxbindings==1.2.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
torchmetrics==0.6.0
torchtext==0.10.0
demucs
gradio>=3.40.1
